Squashed 'lispBM/lispBM/' changes from cd6b6e2a..f0c1ec21
f0c1ec21 fixing type
563be0ef adding memory usage statistics that is collected a the end of each garbage collection round.
135593ea making the repl  vescif_print_ctx clearer
ce8cdaa1 more info about symbols involved in error messages. A binding is looked up if it exists.
3494b2d4 Merge branch 'master' of github.com:svenssonjoel/lispbm
ea1a92db update bear command for recent version!
669c2fc4 added an all contexts iterator
2ab1fd60 update REPL Makefile for SDL 32 and 64bit builds
d00f0472 update REPL makefile for 32 and 64bit SDL builds
9ea2659a update to scan-build script
c34ba6fb added some explanatory comments at the different mark implementaions in heap.c
568fe490 Pointer-reversal GC now works with CALL-CC and Lisp-Arrays. Passes all tests
13138b8e small tweak to cppcheck script
a9627b36 update to cppcheck scripts and a number of refactorings
8e335412 update readme
c06ae021 removing some redundant conditions and reducing scope on a number of variables
66a99c0f reducing scope of a tmp variable in lbm_caar and lbm_cadr
f946c98e fix some tests that were a bit broken and reduce code complexity in print_internal and fundamental_numeq
04ca92ee update to run_cppcheck.sh
8a1ff24b adding scripts for running cppcheck on repl and the libraries the repl uses (most of LBM)
e98cdf33 simplify and clearify share_const_array in lbm_c_interop
9764dafd fix index out of bounds and other issues in repl
64e3f367 added tests that differentiate 32/64bit. mostly a check to make sure one is running the version one thinkgs one is running
4e53e2f0 upping version to differentiate master from 0.27.0
4cf588f9 numbered version 0.27.0
2555dcfc Numbered version 0.27.0
59d5012f fix mistake in tests Makefile
c7ad2521 tweaks to test scripts and test makefile
92f4aeef all 64 bit tests pass. update testing MAkefile to simplify running 32 and 64 bit tests in parallell.
020f4226 corrections related to 64bit testing
ad73aaae default memory size if 10K for the repl
9ccbdad3 profiling now works on the vescif repl
13d5dd49 :prof commands now work in the vescif mode of the repl
ad26c58b added support for the :reset command over the vescif in the REPL
ba4da435 :ctxs now output to vesc_tool if in vesc mode
4b644e25 program upload from vesc-tool to LBM repl seems to work
7a11f0a4 towards vesctcp interoperability, program storage
f94d8822 tweak type in repl help
282c9803 Using a commands_printf_lisp when in vescif mode as the lbm print implementation.
795ba718 You can now set the size of the simulated constants memory for the repl
462034ea update makefile for docs to set an appropriate mem size
97257cd5 close some warnings from buffer.c in repl
ee10423b the LBM linux repl allows selecting lbm_memory size
0dba96fe small tweak to how the repl prints help text upon --help flag
7cb31338 added another example to the repl/examples directory
be1af47a remove unused code from env.c/h
5dfe7813 added a size.txt file to keep track of size (as part of a chibios build) of LBM
819f0f0d consoldating 3 different continuations into one
c02e5f22 remove const-symbol-strings

git-subtree-dir: lispBM/lispBM
git-subtree-split: f0c1ec2124b7d5b14eb370302ac4deb2201e63fe

+++ repl/buffer.c
@@ -137,7 +137,7 @@ void buffer_append_float32_auto(uint8_t* buffer, float number, int32_t *index) {
 		e += 126;
 	}
 
-	uint32_t res = ((e & 0xFF) << 23) | (sig_i & 0x7FFFFF);
+	uint32_t res = ((uint32_t)((e & 0xFF) << 23)) | (sig_i & 0x7FFFFF);
 	if (sig < 0) {
 		res |= 1U << 31;
 	}
@@ -153,26 +153,26 @@ void buffer_append_float64_auto(uint8_t* buffer, double number, int32_t *index) 
 }
 
 int16_t buffer_get_int16(const uint8_t *buffer, int32_t *index) {
-	int16_t res =	((uint16_t) buffer[*index]) << 8 |
-					((uint16_t) buffer[*index + 1]);
-	*index += 2;
-	return res;
+  int16_t res =	(int16_t)((buffer[*index] << 8) |
+                          buffer[*index + 1]);
+  *index += 2;
+  return res;
 }
 
 uint16_t buffer_get_uint16(const uint8_t *buffer, int32_t *index) {
-	uint16_t res = 	((uint16_t) buffer[*index]) << 8 |
-					((uint16_t) buffer[*index + 1]);
-	*index += 2;
-	return res;
+  uint16_t res = 	(uint16_t)((buffer[*index] << 8) |
+                                   buffer[*index + 1]);
+  *index += 2;
+  return res;
 }
 
 int32_t buffer_get_int32(const uint8_t *buffer, int32_t *index) {
-	int32_t res =	((uint32_t) buffer[*index]) << 24 |
-					((uint32_t) buffer[*index + 1]) << 16 |
-					((uint32_t) buffer[*index + 2]) << 8 |
-					((uint32_t) buffer[*index + 3]);
-	*index += 4;
-	return res;
+  int32_t res =	(int32_t)((buffer[*index] << 24) |
+                          (buffer[*index + 1] << 16) |
+                          (buffer[*index + 2] << 8) |
+                          buffer[*index + 3]);
+  *index += 4;
+  return res;
 }
 
 uint32_t buffer_get_uint32(const uint8_t *buffer, int32_t *index) {
@@ -185,16 +185,16 @@ uint32_t buffer_get_uint32(const uint8_t *buffer, int32_t *index) {
 }
 
 int64_t buffer_get_int64(const uint8_t *buffer, int32_t *index) {
-	int64_t res =	((uint64_t) buffer[*index]) << 56 |
-					((uint64_t) buffer[*index + 1]) << 48 |
-					((uint64_t) buffer[*index + 2]) << 40 |
-					((uint64_t) buffer[*index + 3]) << 32 |
-					((uint64_t) buffer[*index + 4]) << 24 |
-					((uint64_t) buffer[*index + 5]) << 16 |
-					((uint64_t) buffer[*index + 6]) << 8 |
-					((uint64_t) buffer[*index + 7]);
-	*index += 8;
-	return res;
+  int64_t res =	(int64_t)(((uint64_t)buffer[*index] << 56) |
+                          ((uint64_t)buffer[*index + 1] << 48) |
+                          ((uint64_t)buffer[*index + 2] << 40) |
+                          ((uint64_t)buffer[*index + 3] << 32) |
+                          ((uint64_t)buffer[*index + 4] << 24) |
+                          ((uint64_t)buffer[*index + 5] << 16) |
+                          ((uint64_t)buffer[*index + 6] << 8) |
+                          ((uint64_t)buffer[*index + 7]));
+  *index += 8;
+  return res;
 }
 
 uint64_t buffer_get_uint64(const uint8_t *buffer, int32_t *index) {
