+++ repl/repl_exts.c
@@ -677,7 +677,7 @@ static void buffer_blast_rgb565(uint8_t *dest, uint8_t *img) {
 
   uint32_t t_pos = 0;
   for (int i = 0; i < num_pix; i ++) {
-    uint16_t pix = (((uint16_t)data[2 * i]) << 8) | ((uint16_t)data[2 * i + 1]);
+    uint16_t pix = (uint16_t)((uint16_t)(data[2 * i] << 8) | ((uint16_t)data[2 * i + 1]));
 
     uint32_t r = (uint32_t)(pix >> 11);
     uint32_t g = (uint32_t)((pix >> 5) & 0x3F);
